requirements.txt
Pillow
python-telegram-bot
flask
imgspy